requirements.txt
fastapi
sqlmodel
asyncpg
aiosqlite
python-jose
httpx
pytest
pytest-asyncio
python-dotenv
uvicorn
pydantic